refactor(TKT-1091): add PMO bootstrapping helpers and migrate pmo/init, refine execution/config to use openWorkspaceDatabase
